ojd_daps/requirements.txt
daps-utils @ git+https://github.com/nestauk/daps_utils@dev#egg=daps-utils
metaflow @ git+https://github.com/jaklinger/metaflow@more_throttle#egg=metaflow
mysql-connector-python==8.0.23
numpy==1.21.1
pytest-timeout==1.4.2
python_dateutil==2.8.1
PyMySQL==1.0.2
PyYAML==5.4
requests>=2.22.0
requests_cache>=0.5.2
SQLAlchemy==1.3.4
bs4==0.0.1
lxml>=4.6.3
cloudscraper>=1.2.48
urllib3>=1.26.5
networkx>=2.6.2
tqdm>=4.62.2
